# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Nervum API
  description: |
    API for Nervum – environment maps SaaS (visualize and manage services, databases,
    queues, and relationships). This backend powers the Nervum web app and can be
    used for integrations.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/users:
    post:
      tags: [Users]
      summary: Create user
      security: [{ cookieAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Users]
      summary: List users
      security: [{ cookieAuth: [] }]
      parameters:
        - name: organization_id
          in: query
          description: Filter by organization (optional)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UUIDPathId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Users]
      summary: Update user
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UUIDPathId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Users]
      summary: Delete user
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UUIDPathId'
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: nervum_session
      description: Session ID (UUID) set by login/register/accept-invitation
  parameters:
    UUIDPathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid request (e.g. validation error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, member]
        organization_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserCreate:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, member]
        organization_id:
          type: string
          format: uuid
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        organization_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [admin, manager, member]
...
security:
  - cookieAuth: []